Move index.d.ts to types/ folder, replaces index.ts file
